Move cancellation off of base worker and add methods to process worker
